Fix merge errors
Removed my old code that made it into master. I literally just took the java files I had from the pause-button branch and ctrl-a ctrl-v nuked MainActivity.java and PlayBackActivity.java, but the app works again. The main error was that there was a textview and an extra constraint layout closing tag in activity_main.xml. The dog gif is always playing in the playback activity and the old pause button is still there.